requirements.txt
django>=1.10,<1.11
django-registration-redux>=1.4,<1.5
django-crispy-forms>=1.6,<1.7
